nfc_utils.py
#!/usr/bin/env python3
# NFC Utilities - Additional utilities for NFC card operations
# Author: AI Assistant

import os
import sys
import time
import random
import binascii
from Crypto.Cipher import DES

class MifareUtils:
    """Utilities for working with MIFARE cards"""

    @staticmethod
    def nested_attack(tag, known_key, known_sector, target_sector):
        """
        Perform a nested authentication attack
        This is a simplified implementation - in a real scenario, this would
        involve capturing nonces and performing cryptographic analysis
        """
        # This is a placeholder for the actual implementation
        # In a real implementation, this would:
        # 1. Authenticate with the known key
        # 2. Capture the nonce from a subsequent authentication attempt
        # 3. Analyze the nonce to recover the key for the target sector

        print(f"Performing nested attack from sector {known_sector} to {target_sector}")
        print(f"Using known key: {known_key.hex()}")

        # Simulate the attack process
        print("Capturing authentication data...")
        time.sleep(1)
        print("Analyzing nonces...")
        time.sleep(2)

        # In a real implementation, this would be the recovered key
        # Here we're just simulating success or failure
        success = random.random() > 0.7

        if success:
            # Generate a random key as our "cracked" key
            cracked_key = bytes([random.randint(0, 255) for _ in range(6)])
            return cracked_key
        else:
            return None

    @staticmethod
    def darkside_attack(tag):
        """
        Simulate a darkside attack on MIFARE Classic
        This attack exploits a weakness in the CRYPTO1 cipher
        """
        print("Performing darkside attack...")
        print("This attack targets a weakness in the CRYPTO1 cipher")

        # Simulate the attack process
        print("Sending specially crafted authentication attempts...")
        time.sleep(2)
        print("Collecting and analyzing responses...")
        time.sleep(2)

        # Simulate success or failure
        success = random.random() > 0.6

        if success:
            # Generate a random key as our "cracked" key
            cracked_key = bytes([random.randint(0, 255) for _ in range(6)])
            return cracked_key
        else:
            return None

class NFCDump:
    """Utilities for dumping and analyzing NFC card data"""

    @staticmethod
    def dump_mifare_classic(tag, keys):
        """Dump all accessible data from a MIFARE Classic card"""
        dump = {}

        # Determine card size
        num_sectors = 16  # Default for MIFARE Classic 1K
        if hasattr(tag, 'size'):
            if tag.size > 1024:
                num_sectors = 40  # MIFARE Classic 4K

        # Try to read all sectors
        for sector in range(num_sectors):
            dump[sector] = {'data': None, 'key_a': None, 'key_b': None}

            # Try authentication with all keys
            for key_type in ['A', 'B']:
                for key in keys:
                    try:
                        if tag.authenticate(sector, key, key_type == 'A'):
                            # Store the working key
                            if key_type == 'A':
                                dump[sector]['key_a'] = key
                            else:
                                dump[sector]['key_b'] = key

                            # Try to read the sector data
                            try:
                                block_start = sector * 4
                                blocks = []
                                for i in range(4):  # 4 blocks per sector
                                    block_data = tag.read(block_start + i)
                                    blocks.append(block_data)
                                dump[sector]['data'] = blocks
                                break
                            except Exception:
                                pass
                    except Exception:
                        continue

        return dump

    @staticmethod
    def save_dump(dump, filename):
        """Save a card dump to a file"""
        with open(filename, 'w') as f:
            for sector, data in dump.items():
                f.write(f"Sector {sector}:\n")

                if data['key_a']:
                    f.write(f"  Key A: {data['key_a'].hex().upper()}\n")
                else:
                    f.write("  Key A: Unknown\n")

                if data['key_b']:
                    f.write(f"  Key B: {data['key_b'].hex().upper()}\n")
                else:
                    f.write("  Key B: Unknown\n")

                if data['data']:
                    f.write("  Data:\n")
                    for i, block in enumerate(data['data']):
                        f.write(f"    Block {sector*4 + i}: {block.hex().upper()}\n")
                else:
                    f.write("  Data: Not accessible\n")

                f.write("\n")

    @staticmethod
    def analyze_dump(dump):
        """Analyze a card dump for common patterns and data"""
        results = {
            'uid': None,
            'card_type': "MIFARE Classic",
            'readable_sectors': 0,
            'total_sectors': len(dump),
            'access_conditions': {},
            'value_blocks': []
        }

        for sector, data in dump.items():
            if data['data']:
                results['readable_sectors'] += 1

                # Extract UID from sector 0, block 0
                if sector == 0 and len(data['data']) > 0:
                    results['uid'] = data['data'][0][:4].hex().upper()

                # Check for value blocks (used for electronic purse)
                for i, block in enumerate(data['data']):
                    block_num = sector * 4 + i

                    # Skip sector trailer blocks (block 3, 7, 11, etc.)
                    if (block_num + 1) % 4 == 0:
                        # This is a sector trailer - extract access conditions
                        if len(block) >= 9:
                            results['access_conditions'][sector] = block[6:9].hex().upper()
                        continue

                    # Check if this might be a value block
                    if len(block) >= 4:
                        value = int.from_bytes(block[0:4], byteorder='little', signed=True)
                        inverted_value = int.from_bytes(block[4:8], byteorder='little', signed=True)
                        if value == ~inverted_value & 0xFFFFFFFF:
                            results['value_blocks'].append({
                                'block': block_num,
                                'value': value
                            })

        return results

class DESFireUtils:
    """Utilities for working with MIFARE DESFire cards"""

    @staticmethod
    def authenticate_desfire(tag, key_id=0, key=b'\x00' * 8):
        """Authenticate to a DESFire card using default DES key"""
        try:
            # This is a simplified implementation
            # In a real scenario, this would involve proper DESFire command handling

            # Select the application (AID)
            aid = b'\x00\x00\x00'  # Master application

            # Authenticate with the key
            cipher = DES.new(key, DES.MODE_CBC, b'\x00' * 8)

            # Simulate authentication process
            print(f"Authenticating to DESFire with key ID {key_id}")
            time.sleep(1)

            # Return simulated success/failure
            return random.random() > 0.5

        except Exception as e:
            print(f"Authentication error: {e}")
            return False

def detect_card_type(tag):
    """Detect and return the card type based on its properties"""
    if hasattr(tag, 'product'):
        if 'MIFARE Classic' in tag.product:
            return "MIFARE Classic"
        elif 'MIFARE Ultralight' in tag.product:
            return "MIFARE Ultralight"
        elif 'MIFARE DESFire' in tag.product:
            return "MIFARE DESFire"
        elif 'FeliCa' in tag.product:
            return "FeliCa"
        else:
            return tag.product

    # Try to determine by other properties
    if hasattr(tag, 'signature') and len(tag.signature) == 32:
        return "MIFARE DESFire or MIFARE Plus"

    return "Unknown"